requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
greenlet
alembic
pgvector
openai
pydantic
pydantic-settings
python-dotenv
tiktoken
python-multipart
sentence-transformers
